Fix asset paths
The problem here is that Webpack `require` doesn't work with Antwar as
expected. It would be better to load all assets through Webpack instead
of copying. Now there's a conflict due to that. Alternatively it could
be set up so that both are allowed to write to the same directories.